+++ core/web/Twig.php
@@ -8,6 +8,8 @@ use Twig_SimpleFunction;
 use Twig_SimpleFilter;
 use Twig_Extension_Debug;
 
+trigger_error('The twig component will be removed in 1.0.0.', E_USER_DEPRECATED);
+
 /**
  * LUYA Twig Component
  *